% Auto-generated: do not edit by hand
\name{dashMathjax}

\alias{dashMathjax}

\title{DashMathjax component}

\description{
DashMathjax is a Dash wrapper for MathJax in Dash (https://github.com/zzish/react-latex).
}

\usage{
dashMathjax(children=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. React element to be rendered inside the GSK Button component.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

